html/semantics/disabled-elements/disable.button-setter.htm

<!DOCTYPE html>
<title>HTMLAnchorElement.text setting</title>
<link rel="author" title="Tim Denney" href="mailto:timdenney@gmail.com">
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/disabled-elements.html#disabled-elements">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>

<button id="test_btn" disabled="disabled" onclick="set_value(this)" >Button</button>

<button id="test_bool_attr1" disabled="" >Button</button>
<button id="test_bool_attr2" disabled>Button</button>
<button id="test_bool_attr3" disabled="false" >Button</button>
<button id="test_bool_attr4" disabled=true >Button</button>
<button id="test_bool_attr5" disabled=false >Button</button>

<input id="test_type_1" type="button" disabled="disabled"/>
<textarea id="test_type_2" type="button" disabled="disabled"></textarea>
<input id="test_type_3" type="password" disabled="disabled"/>
<select id="test_type_4"  disabled="disabled">
  <option>test</option>
</select>
<input id="test_type_5" type="radio" disabled="disabled"/>
<input id="test_type_6" type="checkbox" disabled="disabled"/>

<script>

function set_value(e){
  var btn=document.getElementById(e.id);
  btn.setAttribute("check","fired");
}

function eventFire(item, type){
  if (item!=undefined){
    item.click();
    // var evObj = document.createEvent('Events');
    // evObj.initEvent('click', true, false);
    // item.dispatchEvent(evObj);
  }
}

test(function(){
  var btn=document.getElementById("test_btn");
  assert_equals(btn.getAttribute("disabled"),'disabled');

},'Disable Attribute Exist');

test(function(){
  var btn=document.getElementById("test_btn");
  eventFire(btn,'click');
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Click Event Triggers on Disabled Element');


test(function(){
  var btn=document.getElementById("test_bool_attr1");
  eventFire(btn,'click');
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Boolean Attribute Empty Quotes');

test(function(){
  var btn=document.getElementById("test_bool_attr2");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Boolean Attribute No Value');

test(function(){
  var btn=document.getElementById("test_bool_attr3");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Boolean Attribute False Value');

test(function(){
  var btn=document.getElementById("test_bool_attr4");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Boolean Attribute No Quotes Value (True)');


test(function(){
  var btn=document.getElementById("test_bool_attr5");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Boolean Attribute No Quotes Value (False)');

test(function(){
  var btn=document.getElementById("test_type_1");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Input type textbox should not fire');

test(function(){
  var btn=document.getElementById("test_type_2");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Textarea type should not fire');

test(function(){
  var btn=document.getElementById("test_type_3");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Password type should not fire');

test(function(){
  var btn=document.getElementById("test_type_4");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Select type should not fire');

test(function(){
  var btn=document.getElementById("test_type_5");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Radio type should not fire');

test(function(){
  var btn=document.getElementById("test_type_6");
  assert_not_equals(btn.getAttribute("check"),'fired');

},'Checkbox type should not fire');

</script>
